Implemented chat and made some layout changes.
The tables on the room page (user list and playlist) now scroll as well. It's not done very well though as currently
they're just wrapped in divs with fixed size and scroll bars. It needs to be done so that the table headers stay fixed while
the table body scrolls. Will implement this some time in the future.

Also, the height of the playlist's scroll div now scales automatically with window size.